+++ src/app/admin/components/products/products.component.ts
@@ -4,7 +4,7 @@ import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
 
 import { ProductModel } from '../../../products';
-import { AppState, getProductsData, getProductsLoaded, getProductsLoading, loadProducts } from '../../../core/store';
+import { AppState, getProductsData, getProductsLoaded, getProductsLoading } from '../../../core/store';
 
 @Component({
   selector: 'sh-products',
@@ -21,8 +21,6 @@ export class ProductsComponent implements OnInit {
   ) {}
 
   ngOnInit() {
-    this.store.dispatch(loadProducts());
-
     this.products$ = this.store.select(getProductsData);
     this.loading$ = this.store.select(getProductsLoading);
     this.loaded$ = this.store.select(getProductsLoaded);