+++ src/app/products/guards/index.ts
@@ -1,0 +1,2 @@
+export * from './products-state-preloading.guard';
+export * from './product-exist.guard';